# public/api-documentation/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Laravel Swagger Documentation",
        "contact": {
            "email": "francielefernandes126@gmail.com"
        },
        "version": "1.5.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Realiza login no sistema",
                "description": "Retorna token de autenticação",
                "operationId": "5f90fad45b3bec06f69a5d179ce78a5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "JSON": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email do usuário",
                                        "type": "string",
                                        "example": "email@email.com"
                                    },
                                    "password": {
                                        "description": "Senha do usuário",
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha2023"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação bem sucedida"
                    },
                    "500": {
                        "description": "Erro de servidor interno"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Logout"
                ],
                "summary": "Realiza logout no sistema",
                "description": "Logout do sistema",
                "operationId": "30b14d0552474a26ab0603d4b9691900",
                "responses": {
                    "200": {
                        "description": "Operação bem sucedida"
                    },
                    "500": {
                        "description": "Erro de servidor interno"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/api/favorites": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Favoritar pokemon",
                "operationId": "d426a8b38354581cb4082ac4bc62e22f",
                "requestBody": {
                    "description": "Favoritar Pokemon",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pokemon_id": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "pokemon favoritado com sucesso"
                    },
                    "400": {
                        "description": "Requisição inválida"
                    },
                    "500": {
                        "description": "Erro de servidor interno"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/api/favorites/{id}": {
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Excluir pokemon dos favoritos por ID",
                "operationId": "397d5ad2363742284a8d87c5dc366148",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da pokemon a ser excluída dos favoritos",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "pokemon excluído dos favoritos com sucesso"
                    },
                    "404": {
                        "description": "Favorito não encontrado"
                    },
                    "500": {
                        "description": "Erro de servidor interno"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/api/user-by-name": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Buscar usuários por nome e obter informações sobre Pokémon favoritos",
                "operationId": "0f10a6043746bec769b60c7da4d1a988",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nome para busca",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuários encontrados com sucesso"
                    },
                    "400": {
                        "description": "Erro de validação dos dados"
                    },
                    "404": {
                        "description": "Nenhum usuário encontrado com esse nome"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "description": "Acess token obtido na autenticação",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}